Move pytest base-dir safety helpers to module
Extract base-dir safety and teardown helpers into tests/base_dir_safety.py to avoid importing ambiguous conftest names and centralize logic (issue #1573). Update tests/conftest.py to import and delegate to the new helpers (notably unconfigure_cleanup) and remove duplicated implementations. Adjust tests to reference the new module (tests/setup/test_config_schema.py, tests/test_conftest_base_dir_safety.py, tests/common/test_custom_analysis.py) and add a new test verifying temp base-dir removal. Also tweak a cleanup call to pass max_delete_fraction=1.0 for the small fixture that legitimately removes many files.
