+++ tests/test_conftest_base_dir_safety.py
@@ -1,12 +1,11 @@
 """Tests for pytest base-dir safety helpers."""
 
-import sys
 from pathlib import Path
-from types import ModuleType, SimpleNamespace
+from types import SimpleNamespace
 
 import pytest
 
-import conftest as root_conftest
+from tests import base_dir_safety as bds
 
 
 def _config(
@@ -25,7 +24,7 @@ def _config(
 
 
 def _sentinel(base_dir: Path) -> Path:
-    return base_dir / root_conftest.TEST_ROOT_SENTINEL
+    return base_dir / bds.TEST_ROOT_SENTINEL
 
 
 def _patch_mkdtemp(monkeypatch, tmp_path: Path) -> Path:
@@ -36,15 +35,13 @@ def _patch_mkdtemp(monkeypatch, tmp_path: Path) -> Path:
         tmp_base_dir.mkdir()
         return str(tmp_base_dir)
 
-    monkeypatch.setattr(root_conftest.tempfile, "mkdtemp", _mkdtemp)
+    monkeypatch.setattr(bds.tempfile, "mkdtemp", _mkdtemp)
     return tmp_base_dir
 
 
 def test_resolve_base_dir_rejects_unmarked_cli_path(tmp_path):
     with pytest.raises(pytest.UsageError, match="not marked"):
-        root_conftest._resolve_base_dir(
-            _config(base_dir=str(tmp_path / "shared-data"))
-        )
+        bds._resolve_base_dir(_config(base_dir=str(tmp_path / "shared-data")))
 
 
 def test_resolve_base_dir_accepts_marked_cli_path(tmp_path):
@@ -52,7 +49,7 @@ def test_resolve_base_dir_accepts_marked_cli_path(tmp_path):
     base_dir.mkdir()
     _sentinel(base_dir).write_text("test sandbox\n")
 
-    resolved, tmp_base_dir = root_conftest._resolve_base_dir(
+    resolved, tmp_base_dir = bds._resolve_base_dir(
         _config(base_dir=str(base_dir))
     )
 
@@ -66,7 +63,7 @@ def test_resolve_base_dir_ignores_env_without_flag(monkeypatch, tmp_path):
     tmp_base_dir = _patch_mkdtemp(monkeypatch, tmp_path)
 
     with pytest.warns(RuntimeWarning, match="Ignoring SPYGLASS_BASE_DIR"):
-        resolved, tmp_dir = root_conftest._resolve_base_dir(_config())
+        resolved, tmp_dir = bds._resolve_base_dir(_config())
 
     assert Path(resolved) == tmp_base_dir
     assert tmp_dir == str(tmp_base_dir)
@@ -79,7 +76,7 @@ def test_resolve_base_dir_uses_marked_env_with_flag(monkeypatch, tmp_path):
     _sentinel(env_base).write_text("test sandbox\n")
     monkeypatch.setenv("SPYGLASS_BASE_DIR", str(env_base))
 
-    resolved, tmp_base_dir = root_conftest._resolve_base_dir(
+    resolved, tmp_base_dir = bds._resolve_base_dir(
         _config(use_env_base_dir=True)
     )
 
@@ -92,7 +89,7 @@ def test_resolve_base_dir_rejects_unmarked_env_with_flag(monkeypatch, tmp_path):
     monkeypatch.setenv("SPYGLASS_BASE_DIR", str(env_base))
 
     with pytest.raises(pytest.UsageError, match="not marked"):
-        root_conftest._resolve_base_dir(_config(use_env_base_dir=True))
+        bds._resolve_base_dir(_config(use_env_base_dir=True))
 
 
 def test_resolve_base_dir_warns_when_env_flag_has_no_env(monkeypatch, tmp_path):
@@ -100,7 +97,7 @@ def test_resolve_base_dir_warns_when_env_flag_has_no_env(monkeypatch, tmp_path):
     tmp_base_dir = _patch_mkdtemp(monkeypatch, tmp_path)
 
     with pytest.warns(RuntimeWarning, match="SPYGLASS_BASE_DIR is not set"):
-        resolved, tmp_dir = root_conftest._resolve_base_dir(
+        resolved, tmp_dir = bds._resolve_base_dir(
             _config(use_env_base_dir=True)
         )
 
@@ -111,7 +108,7 @@ def test_resolve_base_dir_warns_when_env_flag_has_no_env(monkeypatch, tmp_path):
 
 def test_resolve_base_dir_rejects_no_teardown_without_persistent_base():
     with pytest.raises(pytest.UsageError, match="--no-teardown requires"):
-        root_conftest._resolve_base_dir(_config(no_teardown=True))
+        bds._resolve_base_dir(_config(no_teardown=True))
 
 
 class _FakeServer:
@@ -122,9 +119,7 @@ class _FakeServer:
         self.stopped = True
 
 
-def test_unconfigure_does_not_clean_user_supplied_base_dir(
-    monkeypatch, tmp_path
-):
+def test_unconfigure_does_not_clean_user_supplied_base_dir(tmp_path):
     base_dir = tmp_path / "persistent-base"
     analysis_file = base_dir / "analysis" / "old-result.nwb"
     tmp_subdir_file = base_dir / "tmp" / "old-artifact.txt"
@@ -132,22 +127,32 @@ def test_unconfigure_does_not_clean_user_supplied_base_dir(
     tmp_subdir_file.parent.mkdir(parents=True)
     analysis_file.write_text("analysis")
     tmp_subdir_file.write_text("artifact")
-    _sentinel(base_dir).write_text("test sandbox\n")
 
     fake_server = _FakeServer()
-    fake_nwb_helper = ModuleType("spyglass.utils.nwb_helper_fn")
-    fake_nwb_helper.close_nwb_files = lambda: None
-
-    monkeypatch.setitem(root_conftest.__dict__, "SERVER", fake_server)
-    monkeypatch.setitem(root_conftest.__dict__, "TMP_BASE_DIR", None)
-    monkeypatch.setitem(root_conftest.__dict__, "BASE_DIR", base_dir)
-    monkeypatch.setitem(root_conftest.__dict__, "TEARDOWN", True)
-    monkeypatch.setitem(
-        sys.modules, "spyglass.utils.nwb_helper_fn", fake_nwb_helper
+    bds.unconfigure_cleanup(
+        server=fake_server,
+        tmp_base_dir=None,
+        teardown=True,
+        close_nwb_files=lambda: None,
     )
-
-    root_conftest.pytest_unconfigure(_config())
 
     assert fake_server.stopped
     assert analysis_file.exists()
     assert tmp_subdir_file.exists()
+
+
+def test_unconfigure_removes_temp_base_dir(tmp_path):
+    tmp_base_dir = tmp_path / "spyglass_test_session"
+    tmp_base_dir.mkdir()
+    (tmp_base_dir / "scratch.nwb").write_text("scratch")
+
+    fake_server = _FakeServer()
+    bds.unconfigure_cleanup(
+        server=fake_server,
+        tmp_base_dir=str(tmp_base_dir),
+        teardown=True,
+        close_nwb_files=lambda: None,
+    )
+
+    assert fake_server.stopped
+    assert not tmp_base_dir.exists()
